upgrade: gov upgrade handler changes (#186)
* update readme

* go.mod: bump up wasmd

* updated expedited proposal min deposit

* fix lint

* fix lint

* update globalfee params

+++ x/globalfee/types/params.go
@@ -9,14 +9,14 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-var maxTotalBypassMinFeeMsgGasUsage = 2_000_000
+var maxTotalBypassMinFeeMsgGasUsage = uint64(2_000_000)
 
 // DefaultParams returns default parameters
 func DefaultParams() Params {
 	return Params{
 		MinimumGasPrices:                sdk.DecCoins(nil),
 		BypassMinFeeMsgTypes:            []string{},
-		MaxTotalBypassMinFeeMsgGasUsage: uint64(maxTotalBypassMinFeeMsgGasUsage),
+		MaxTotalBypassMinFeeMsgGasUsage: maxTotalBypassMinFeeMsgGasUsage,
 	}
 }
